% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_all_before_RNAseq.R
\name{CheckAllParameter}
\alias{CheckAllParameter}
\title{Check everything before calling 'RNAseqPipelineCMD'}
\usage{
CheckAllParameter(path.prefix = "NOT_SET_YET",
  input.path.prefix = "NOT_SET_YET", gene.name = "NO_DATA",
  sample.pattern = "NO_DATA")
}
\description{
Check everything before calling 'RNAseqPipelineCMD'
}
