% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tool_command.R
\name{CheckBallgown}
\alias{CheckBallgown}
\title{Check Ballgown
Check whether Ballgown is installed on the workstation}
\usage{
CheckBallgown(print = TRUE)
}
\description{
Check Ballgown
Check whether Ballgown is installed on the workstation
}
