% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_precheck.R
\name{CheckGffcompare}
\alias{CheckGffcompare}
\title{Check Gffcompare
Check whether Gffcompare is installed on the workstation}
\usage{
CheckGffcompare()
}
\description{
Check Gffcompare
Check whether Gffcompare is installed on the workstation
}
