% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_files_directory.R
\name{CheckInputDir}
\alias{CheckInputDir}
\title{Check input files directory}
\usage{
CheckInputDir(gene_name = "NO_DATA", abs_input_dir = "NOT_SET_YET",
  print = TRUE)
}
\description{
Check input files directory
}
