% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_files_directory.R
\name{CheckInputDirFiles}
\alias{CheckInputDirFiles}
\title{Check input files directory}
\usage{
CheckInputDirFiles(input.path.prefix = "NOT_SET_YET", gene.name = "NO_DATA",
  sample.pattern = "NO_DATA", print = TRUE)
}
\description{
Check input files directory
}
