% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_path_setting.R
\name{CheckPrefixPath}
\alias{CheckPrefixPath}
\title{Checking the absolute path}
\usage{
CheckPrefixPath(path.prefix = pkg.global.path.prefix$data_path,
  print = TRUE)
}
\description{
Checking the absolute path
}
