% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisat2_align_default.R
\name{CheckSampleGenesFiles}
\alias{CheckSampleGenesFiles}
\title{check 'gene_data' and subdirectory files exit}
\usage{
CheckSampleGenesFiles(gene_name = "NO_DATA")
}
\description{
check 'gene_data' and subdirectory files exit
}
