% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tool_command.R
\name{CheckSamtools}
\alias{CheckSamtools}
\title{Check Samtools
Check whether Samtools is installed on the workstation}
\usage{
CheckSamtools(print = TRUE)
}
\description{
Check Samtools
Check whether Samtools is installed on the workstation
}
