% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_precheck.R
\name{CheckStringTie}
\alias{CheckStringTie}
\title{Check StringTie
Check whether StringTie is installed on the workstation}
\usage{
CheckStringTie()
}
\description{
Check StringTie
Check whether StringTie is installed on the workstation
}
