% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_files_directory.R
\name{CopyInputDir}
\alias{CopyInputDir}
\title{Copy input files directory}
\usage{
CopyInputDir(input.path.prefix = "NOT_SET_YET", gene.name = "NO_DATA",
  sample.pattern = "NO_DATA")
}
\description{
Copy input files directory
}
