% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gffcompare_related.R
\name{GffcompareRefSample}
\alias{GffcompareRefSample}
\title{Examine how the transcripts compare with the reference annotation}
\usage{
GffcompareRefSample(gene_name = "NO_DATA", sample_prefix = "NO_DATA")
}
\description{
Examine how the transcripts compare with the reference annotation
}
