% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisat2_related.R
\name{Hisat2AlignmentDefault}
\alias{Hisat2AlignmentDefault}
\title{hisat2 alignment default}
\usage{
Hisat2AlignmentDefault(gene_name = "NO_DATA", sample_prefix = "NO_DATA")
}
\description{
hisat2 alignment default
}
