% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_install_tools.R
\name{InstallToolsCMD}
\alias{InstallToolsCMD}
\title{install all the tools and create a binary library.}
\usage{
InstallToolsCMD(path.prefix = "NOT_SET_YET",
  input.path.prefix = "NOT_SET_YET", gene.name = "NO_DATA",
  sample.pattern = "NO_DATA")
}
\description{
install all the tools and create a binary library.
}
