% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_pipeline.R
\name{RNAseqPipeline}
\alias{RNAseqPipeline}
\title{rna seq pipline}
\usage{
RNAseqPipeline(path.prefix = home.path, input.files.path = "NOT_SET_YET",
  gene_name = "NO_DATA")
}
\description{
rna seq pipline
}
