% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_pipline.R
\name{RNAseqPipelineCMD}
\alias{RNAseqPipelineCMD}
\title{Create 'RNAseqPipline.R' that user can}
\usage{
RNAseqPipelineCMD(path.prefix = "NOT_SET_YET",
  input.files.path = "NOT_SET_YET", gene_name = "NO_DATA",
  sample_prefix = "NO_DATA")
}
\description{
Create 'RNAseqPipline.R' that user can
}
