% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_path_setting.R
\name{SetPrefixPath}
\alias{SetPrefixPath}
\title{Choose the location where 'gene_data' and 'RNAseq_bin' will be installed}
\usage{
SetPrefixPath(path.prefix = "NOT_SET_YET", print = TRUE)
}
\description{
Choose the location where 'gene_data' and 'RNAseq_bin' will be installed
}
