% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisat2_samtools_stringtie_ballgown_related.R,
%   R/stringtie_related.R
\name{StringTieAssemble}
\alias{StringTieAssemble}
\alias{StringTieAssemble}
\title{stringtie assemble and quantify expressed genes and transcripts}
\usage{
StringTieAssemble(gene.name = "NO_DATA", sample.pattern = "NO_DATA",
  num.parallel.threads = 8)

StringTieAssemble(gene.name = "NO_DATA", sample.pattern = "NO_DATA",
  num.parallel.threads = 8)
}
\description{
stringtie assemble and quantify expressed genes and transcripts

stringtie assemble and quantify expressed genes and transcripts
}
