% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisat2_samtools_stringtie_ballgown_related.R,
%   R/stringtie_related.R
\name{StringTieMergeTrans}
\alias{StringTieMergeTrans}
\alias{StringTieMergeTrans}
\title{stringtie merge transcripts from all samples}
\usage{
StringTieMergeTrans(gene.name = "NO_DATA", sample.pattern = "NO_DATA",
  num.parallel.threads = 8)

StringTieMergeTrans(gene.name = "NO_DATA", sample.pattern = "NO_DATA",
  num.parallel.threads = 8)
}
\description{
stringtie merge transcripts from all samples

stringtie merge transcripts from all samples
}
