% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringtie_related.R
\name{StringTieToBallgown}
\alias{StringTieToBallgown}
\title{stringtie estimate transcript abundances and create table counts for Ballgown}
\usage{
StringTieToBallgown(gene_name = "NO_DATA", sample_prefix = "NO_DATA")
}
\description{
stringtie estimate transcript abundances and create table counts for Ballgown
}
