% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tool_command.R
\name{CheckHisat2}
\alias{CheckHisat2}
\title{Check 'hisat2'
Check whether 'hisat2' is installed on the workstation}
\usage{
CheckHisat2(print = TRUE)
}
\description{
Check 'hisat2'
Check whether 'hisat2' is installed on the workstation
}
