% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_files_directory.R
\name{CheckInputDirFiles}
\alias{CheckInputDirFiles}
\title{Check input files directory}
\usage{
CheckInputDirFiles(gene_name = "NO_DATA", abs.input.dir = "NOT_SET_YET",
  print = TRUE)
}
\description{
Check input files directory
}
