% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_precheck.R
\name{CheckSamtools}
\alias{CheckSamtools}
\title{Check Samtools
Check whether Samtools is installed on the workstation}
\usage{
CheckSamtools()
}
\description{
Check Samtools
Check whether Samtools is installed on the workstation
}
