% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tool_command.R
\name{CheckStringTie}
\alias{CheckStringTie}
\title{Check s'tringtie'
Check whether 'stringtie' is installed on the workstation}
\usage{
CheckStringTie(print = TRUE)
}
\description{
Check s'tringtie'
Check whether 'stringtie' is installed on the workstation
}
