% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_installtool_rnaseqpipline.R
\name{CopyInputDir}
\alias{CopyInputDir}
\title{Copy input files directory}
\usage{
CopyInputDir(input.path.prefix = "NOT_SET_YET", gene.name = "NO_DATA",
  sample.pattern = "NO_DATA", optional = pkg.global.ht2$logic)
}
\description{
Copy input files directory
}
