% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisat2_samtools_stringtie_ballgown_related.R
\name{Hisat2AlignmentDefault}
\alias{Hisat2AlignmentDefault}
\title{hisat2 alignment default}
\usage{
Hisat2AlignmentDefault(gene.name = "NO_DATA", sample.pattern = "NO_DATA",
  num.parallel.threads = 8)
}
\description{
hisat2 alignment default
}
