% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sample_genes_files.R
\name{ProgressGenesFiles}
\alias{ProgressGenesFiles}
\title{check 'gene_data' and subdirectory files exit}
\usage{
ProgressGenesFiles(gene.name = "NO_DATA", sample.pattern = "NO_DATA",
  print = TRUE)
}
\description{
check 'gene_data' and subdirectory files exit
}
