% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samtools_related.R
\name{SamtoolsToBam}
\alias{SamtoolsToBam}
\title{use 'samtools' to sort and convert the SAM files to BAM}
\usage{
SamtoolsToBam(gene.name = "NO_DATA", sample.pattern = "NO_DATA")
}
\description{
use 'samtools' to sort and convert the SAM files to BAM
}
