% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringtie_related.R
\name{StringTieAssemble}
\alias{StringTieAssemble}
\title{stringtie assemble and quantify expressed genes and transcripts}
\usage{
StringTieAssemble(gene.name = "NO_DATA", sample.pattern = "NO_DATA")
}
\description{
stringtie assemble and quantify expressed genes and transcripts
}
