% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringtie_related.R
\name{StringTieAssemble}
\alias{StringTieAssemble}
\title{stringtie assemble and quantify expressed genes and transcripts}
\usage{
StringTieAssemble(gene_name = "NO_DATA")
}
\description{
stringtie assemble and quantify expressed genes and transcripts
}
