% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringtie_related.R
\name{StringTieMergeTrans}
\alias{StringTieMergeTrans}
\title{stringtie merge transcripts from all samples}
\usage{
StringTieMergeTrans(gene.name = "NO_DATA", sample.pattern = "NO_DATA")
}
\description{
stringtie merge transcripts from all samples
}
